// clients/js/src/generated/accounts/coachAuthority.ts
export type CoachAuthority = {
  discriminator: ReadonlyUint8Array;
  runId: Array<number>;
  profileId: Array<number>;
  originalWallet: Address;
  currentWallet: Address;
  recoveryAuthority: Address;
  authorityEpoch: bigint;
  eventSequence: bigint;
  bump: number;
  reserved: Array<number>;
};