// clients/js/src/generated/types/contributionStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum ContributionStatus {
  Funded,
  Refunded,
}

export type ContributionStatusArgs = ContributionStatus;

export function getContributionStatusEncoder(): FixedSizeEncoder<ContributionStatusArgs> {
  return getEnumEncoder(ContributionStatus);
}

export function getContributionStatusDecoder(): FixedSizeDecoder<ContributionStatus> {
  return getEnumDecoder(ContributionStatus);
}

export function getContributionStatusCodec(): FixedSizeCodec<
  ContributionStatusArgs,
  ContributionStatus
> {
  return combineCodec(
    getContributionStatusEncoder(),
    getContributionStatusDecoder(),
  );
}